requirements.txt
torch
torchvision
torchaudio
openai-clip
Pillow
ffmpeg-python
pytz
suntime
piexif
requests
